Initiate migration to modular code
Added models folder where all the various models will be defined
There is a class BaseModel which should be inherited by every model. It
consists of various functions for model saving, evaluation, training and
input loading.

This will be a continous process and it will take some time to migrate
all code to this format.